Move `tcx`-dependent code out of deconstruct_pat
