+++ compiler/rustc_mir_build/src/thir/pattern/mod.rs
@@ -11,22 +11,24 @@ use crate::thir::util::UserAnnotatedTyHelpers;
 use deconstruct_pat::{Constructor, DeconstructedPat, Fields, IntRange, Slice, SliceKind};
 use usefulness::MatchCheckCtxt;
 
+use rustc_data_structures::captures::Captures;
 use rustc_errors::struct_span_err;
 use rustc_hir as hir;
 use rustc_hir::def::{CtorOf, DefKind, Res};
 use rustc_hir::pat_util::EnumerateAndAdjustIterator;
 use rustc_hir::RangeEnd;
 use rustc_index::vec::Idx;
+use rustc_middle::middle::stability::EvalResult;
 use rustc_middle::mir::interpret::{
     get_slice_bytes, ConstValue, ErrorHandled, LitToConstError, LitToConstInput, Scalar,
 };
 use rustc_middle::mir::{BorrowKind, Field, Mutability};
 use rustc_middle::thir::{Ascription, BindingMode, FieldPat, Pat, PatKind, PatRange, PatTyProj};
-use rustc_middle::ty::{self, ConstKind, DefIdTree, Ty, TyCtxt};
+use rustc_middle::ty::{self, ConstKind, DefIdTree, Ty, TyCtxt, VariantDef};
 use rustc_span::{Span, DUMMY_SP};
 use rustc_target::abi::Size;
 
-use smallvec::SmallVec;
+use smallvec::{smallvec, SmallVec};
 use std::cmp::Ordering;
 
 #[derive(Clone, Debug)]
@@ -743,12 +745,12 @@ fn pat_to_deconstructed<'p, 'tcx>(
                     // For each field in the variant, we store the relevant index into `self.fields` if any.
                     let mut field_id_to_id: Vec<Option<usize>> =
                         (0..variant.fields.len()).map(|_| None).collect();
-                    let tys = Fields::list_variant_nonhidden_fields(cx, pat.ty, variant)
-                        .enumerate()
-                        .map(|(i, (field, ty))| {
+                    let tys = cx.list_variant_nonhidden_fields(pat.ty, variant).enumerate().map(
+                        |(i, (field, ty))| {
                             field_id_to_id[field.index()] = Some(i);
                             ty
-                        });
+                        },
+                    );
                     let mut wilds: SmallVec<[_; 2]> = tys.map(DeconstructedPat::wildcard).collect();
                     for pat in subpatterns {
                         if let Some(i) = field_id_to_id[pat.field.index()] {
@@ -893,7 +895,8 @@ fn deconstructed_to_pat<'p, 'tcx>(
             ty::Adt(adt_def, substs) => {
                 let variant_index = pat.ctor().variant_index_for_adt(adt_def);
                 let variant = &adt_def.variants[variant_index];
-                let subpatterns = Fields::list_variant_nonhidden_fields(cx, pat.ty(), variant)
+                let subpatterns = cx
+                    .list_variant_nonhidden_fields(pat.ty(), variant)
                     .zip(subpatterns)
                     .map(|((field, _ty), pattern)| FieldPat { field, pattern })
                     .collect();
@@ -971,4 +974,197 @@ fn deconstructed_to_pat<'p, 'tcx>(
     };
 
     Pat { ty: pat.ty(), span: DUMMY_SP, kind: Box::new(kind) }
+}
+
+impl<'p, 'tcx> MatchCheckCtxt<'p, 'tcx> {
+    fn is_uninhabited(&self, ty: Ty<'tcx>) -> bool {
+        if self.tcx.features().exhaustive_patterns {
+            self.tcx.is_ty_uninhabited_from(self.module, ty, self.param_env)
+        } else {
+            false
+        }
+    }
+
+    /// Returns whether the given type is an enum from another crate declared `#[non_exhaustive]`.
+    fn is_foreign_non_exhaustive_enum(&self, ty: Ty<'tcx>) -> bool {
+        match ty.kind() {
+            ty::Adt(def, ..) => {
+                def.is_enum() && def.is_variant_list_non_exhaustive() && !def.did.is_local()
+            }
+            _ => false,
+        }
+    }
+
+    /// List the set of possible constructors for the type `ty`. For numbers, arrays and slices we use
+    /// ranges and variable-length slices when appropriate.
+    ///
+    /// If the `exhaustive_patterns` feature is enabled, we make sure to omit constructors that are
+    /// statically impossible. E.g., for `Option<!>`, we do not include `Some(_)` in the returned list
+    /// of constructors.
+    /// Even if `exhaustive_patterns` is disabled, we also omit constructors for empty enums and the
+    /// never type if we are at the top-level in a match, because we want to allow empty matches to be
+    /// valid for those.
+    /// Invariant: this is only empty when the type is uninhabited, and the reverse holds if
+    /// `exhaustive_patterns` is enabled.
+    fn list_constructors_for_type(
+        &self,
+        ty: Ty<'tcx>,
+        is_top_level_pattern: bool,
+    ) -> SmallVec<[Constructor<'tcx>; 1]> {
+        use Constructor::*;
+        use SliceKind::*;
+        let cx = self;
+        debug!("list_constructors_for_type({:?})", ty);
+        let make_range = |size, start, end| {
+            IntRange(self::IntRange::from_bits(ty, size, start, end, &RangeEnd::Included))
+        };
+        match ty.kind() {
+            _ if cx.is_uninhabited(ty) => smallvec![],
+            ty::Bool => smallvec![Bool(false), Bool(true)],
+            ty::Char => {
+                let size = number_type_size(cx.tcx, ty);
+                smallvec![
+                    // The valid Unicode Scalar Value ranges.
+                    make_range(size, '\u{0000}' as u128, '\u{D7FF}' as u128),
+                    make_range(size, '\u{E000}' as u128, '\u{10FFFF}' as u128),
+                ]
+            }
+            ty::Int(_) | ty::Uint(_)
+                if ty.is_ptr_sized_integral()
+                    && !cx.tcx.features().precise_pointer_size_matching =>
+            {
+                // `usize`/`isize` are not allowed to be matched exhaustively unless the
+                // `precise_pointer_size_matching` feature is enabled. So we treat those types like
+                // `#[non_exhaustive]` enums by returning a special unmatcheable constructor.
+                smallvec![NonExhaustive]
+            }
+            &ty::Int(_) => {
+                let size = number_type_size(cx.tcx, ty);
+                let bits = size.bits() as u128;
+                let min = 1u128 << (bits - 1);
+                let max = min - 1;
+                smallvec![make_range(size, min, max)]
+            }
+            &ty::Uint(_) => {
+                let size = number_type_size(cx.tcx, ty);
+                let max = size.truncate(u128::MAX);
+                smallvec![make_range(size, 0, max)]
+            }
+            ty::Adt(def, substs) if def.is_enum() => {
+                let exhaustive_patterns_feature_enabled = cx.tcx.features().exhaustive_patterns;
+                let mut ctors: SmallVec<[_; 1]> = def
+                    .variants
+                    .iter_enumerated()
+                    .filter(|(_, v)| {
+                        // Filter variants that depend on a disabled unstable feature.
+                        let is_enabled = !matches!(
+                            cx.tcx.eval_stability(v.def_id, None, DUMMY_SP, None),
+                            EvalResult::Deny { .. }
+                        );
+                        // If `exhaustive_patterns` is enabled, we exclude variants known to be
+                        // uninhabited.
+                        let is_uninhabited = exhaustive_patterns_feature_enabled
+                            && v.uninhabited_from(cx.tcx, substs, def.adt_kind(), cx.param_env)
+                                .contains(cx.tcx, cx.module);
+                        is_enabled && !is_uninhabited
+                    })
+                    .map(|(idx, _)| Variant(idx))
+                    .collect();
+
+                // If the enum is declared as `#[non_exhaustive]`, we treat it as if it had an
+                // additional "unknown" constructor.
+                // There is no point in enumerating all possible variants, because the user can't
+                // actually match against them all themselves. So we always return only the fictitious
+                // constructor.
+                // E.g., in an example like:
+                //
+                // ```
+                //     let err: io::ErrorKind = ...;
+                //     match err {
+                //         io::ErrorKind::NotFound => {},
+                //     }
+                // ```
+                //
+                // we don't want to show every possible IO error, but instead have only `_` as the
+                // witness.
+                let is_declared_nonexhaustive = cx.is_foreign_non_exhaustive_enum(ty);
+
+                // If `exhaustive_patterns` is disabled and our scrutinee is an empty enum, we treat it
+                // as though it had an "unknown" constructor to avoid exposing its emptiness. The
+                // exception is if the pattern is at the top level, because we want empty matches to be
+                // considered exhaustive.
+                let is_secretly_empty = def.variants.is_empty()
+                    && !exhaustive_patterns_feature_enabled
+                    && !is_top_level_pattern;
+
+                if is_declared_nonexhaustive || is_secretly_empty {
+                    ctors.push(NonExhaustive);
+                }
+                ctors
+            }
+            // The only legal non-wildcard patterns of type `Box` are box patterns, so we emit
+            // that.
+            ty::Adt(def, substs) if def.is_box() => smallvec![BoxPat(substs.type_at(0))],
+            ty::Adt(..) => smallvec![Single],
+            ty::Ref(_, ty, _) => smallvec![Ref(*ty)],
+            ty::Tuple(fs) => smallvec![Tuple(*fs)],
+            ty::Array(sub_ty, len) => {
+                match len.try_eval_usize(cx.tcx, cx.param_env) {
+                    Some(len) => {
+                        // The uninhabited case is already handled.
+                        smallvec![Slice(self::Slice::new(Some(len as usize), VarLen(0, 0)))]
+                    }
+                    None => {
+                        // Treat arrays of a constant but unknown length like slices.
+                        let kind =
+                            if cx.is_uninhabited(sub_ty) { FixedLen(0) } else { VarLen(0, 0) };
+                        smallvec![Slice(self::Slice::new(None, kind))]
+                    }
+                }
+            }
+            ty::Slice(sub_ty) => {
+                let kind = if cx.is_uninhabited(sub_ty) { FixedLen(0) } else { VarLen(0, 0) };
+                smallvec![Slice(self::Slice::new(None, kind))]
+            }
+            // If `exhaustive_patterns` is disabled and our scrutinee is the never type, we don't
+            // expose its emptiness and return `NonExhaustive` below. The exception is if the
+            // pattern is at the top level, because we want empty matches to be considered
+            // exhaustive.
+            ty::Never if is_top_level_pattern => smallvec![],
+            // This type is one for which we cannot list constructors, like `str` or `f64`, or is
+            // uninhabited but `exhaustive_patterns` is disabled.
+            _ => smallvec![NonExhaustive],
+        }
+    }
+
+    /// List the fields for this variant.
+    /// In the cases of either a `#[non_exhaustive]` field list or a non-public field, we hide
+    /// uninhabited fields in order not to reveal the uninhabitedness of the whole variant.
+    fn list_variant_nonhidden_fields<'a>(
+        &'a self,
+        ty: Ty<'tcx>,
+        variant: &'a VariantDef,
+    ) -> impl Iterator<Item = (Field, Ty<'tcx>)> + Captures<'a> + Captures<'p> {
+        let cx = self;
+        let (adt, substs) = match ty.kind() {
+            ty::Adt(adt, substs) => (adt, substs),
+            _ => bug!(),
+        };
+        // Whether we must not match the fields of this variant exhaustively.
+        let is_non_exhaustive = variant.is_field_list_non_exhaustive() && !adt.did.is_local();
+
+        variant.fields.iter().enumerate().filter_map(move |(i, field)| {
+            let ty = field.ty(cx.tcx, substs);
+            // `field.ty()` doesn't normalize after substituting.
+            let ty = cx.tcx.normalize_erasing_regions(cx.param_env, ty);
+            let is_visible = adt.is_enum() || field.vis.is_accessible_from(cx.module, cx.tcx);
+            let is_uninhabited = cx.is_uninhabited(ty);
+
+            if is_uninhabited && (!is_visible || is_non_exhaustive) {
+                None
+            } else {
+                Some((Field::new(i), ty))
+            }
+        })
+    }
 }
